+++ bare-metal/fedora-atomic/kubernetes/outputs.tf
@@ -1,3 +1,0 @@
-output "kubeconfig-admin" {
-  value = "${module.bootkube.kubeconfig-admin}"
-}